Rendering of content for admins and vendors based on options from sidebar menu, organised folders and added newly created roles enum on more files
